+++ CHANGELOG.md
@@ -7,6 +7,12 @@
   * `bgTheme`: 'warning', 'default', 'error', 'info', 'new', 'success', 'help', or 'maintenance'
   * `bgSize`: 'small' (default), 'medium', or 'large' - only modifies overall icon size if `bgShape` or `bgTheme` is passed.
 
+* `Form`: Two new props have been added, `saveButtonProps` and `cancelButtonProps`, to apply props to Form buttons.
+
+## CSS Changes
+
+* Added `$beta` orange color variable
+
 # 0.30.0
 
 ## SettingsRow Component